move proxy logic to TcpClientAdapter
